requirements.txt
Flask>=2.3
paho-mqtt>=1.6
gpiozero>=1.6
pygame>=2.5
PyYAML>=6.0
Jinja2>=3.1
python-dotenv>=1.0
waitress>=2.1
typing-extensions>=4.9
schedule>=1.2
